p11-platform/services/mcp-servers/meta-ads/requirements.txt
# MCP SDK
mcp>=0.9.0

# HTTP client for Graph API
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
supabase>=2.0.0

